2.2.0.006 snapshot (#84)
* Update INFO

Add file location wizard and UI preparation

* Create Folder

WIZARD_UIFILES

* Add files via upload

Add file location wizard and UI preparation

* Update installer.sh

Fixes backup strategy by selecting another location as before

* Add files via upload

Change wizard description and add html unicode

* Update INFO

Add a install logfile (/var/log/openHAB2-install.log)

* Update installer.sh

Add a install logfile (/var/log/openHAB2-install.log)
Fixes the conf and addons restore

+++ scripts/installer.sh
@@ -3,6 +3,14 @@
 #--------openHAB2 installer script
 #--------package based on work from pcloadletter.co.uk
 
+LOG="/var/log/${SYNOPKG_PKGNAME}-install.log"
+# Delete Log if older than 1 day
+find ${LOG} -mtime +1 -type f -delete
+echo "#### S T A R T  -  o p e n H A B  S P K ####" >>$LOG
+echo "$(date +%Y-%m-%d:%H:%M:%S)" >>$LOG
+echo "" >>$LOG
+
+echo "Set instance variables..." >>$LOG
 DOWNLOAD_PATH="https://openhab.ci.cloudbees.com/job/openHAB-Distribution/lastSuccessfulBuild/artifact/distributions/openhab/target"
 DOWNLOAD_FILE1="openhab-2.2.0-SNAPSHOT.tar.gz"
 
@@ -22,73 +30,96 @@ source /etc/profile
 
 TEMP_FOLDER="$(find / -maxdepth 2 -name '@tmp' | head -n 1)"
 PRIMARY_VOLUME="$(echo ${TEMP_FOLDER} | grep -oP '^/[^/]+')"
-PUBLIC_FOLDER="$(synoshare --get public | grep -oP 'Path.+\[\K[^]]+')"
 
-PUBLIC_OH_FOLDERS_EXISTS=no
-PUBLIC_CONF="${PUBLIC_FOLDER}/openHAB2/conf"
-PUBLIC_ADDONS="${PUBLIC_FOLDER}/openHAB2/addons"
+if [ "${pkgwizard_public_std}" == "true" ]; then
+  SHARE_FOLDER="$(synoshare --get public | grep -oP 'Path.+\[\K[^]]+')"
+  OH_FOLDER="${SHARE_FOLDER}/${SYNOPKG_PKGNAME}"
+elif [ "${pkgwizard_public_shome}"  == "true" ]; then
+  SHARE_FOLDER="$(synoshare --get smarthome | grep -oP 'Path.+\[\K[^]]+')"
+  OH_FOLDER="${SHARE_FOLDER}/${SYNOPKG_PKGNAME}"
+else
+  SHARE_FOLDER="/var/services/homes"
+  OH_FOLDER="${SHARE_FOLDER}/${DAEMON_USER}"
+fi
+
+OH_FOLDERS_EXISTS=no
+OH_CONF="${OH_FOLDER}/conf"
+OH_ADDONS="${OH_FOLDER}/addons"
 TIMESTAMP="$(date +%Y%m)"
 BACKUP_FOLDER="${SYNOPKG_PKGDEST}-backup-$TIMESTAMP"
 
+echo "  primary: ${PRIMARY_VOLUME}" >>$LOG
+echo "  share:   ${SHARE_FOLDER}" >>$LOG
+echo "  oh:      ${OH_FOLDER}" >>$LOG
+echo "  backup:  ${BACKUP_FOLDER}" >>$LOG
+echo "done" >>$LOG
+
 preinst ()
 {
+  echo "Start preinst..." >>$LOG
   # Is Java properly installed?
   if [[ -z "${JAVA_HOME}" || ! -f "${JAVA_HOME}/bin/java" ]]; then
-    echo "Java is not installed or not properly configured."
-    echo "Download and install as described on http://wp.me/pVshC-z5"
-    echo "The Synology provided Java may not work with OpenHAB."
+    echo "  ERROR:" >>$LOG
+    echo "  Java is not installed or not properly configured." >>$LOG
+    echo "  Download and install as described on http://wp.me/pVshC-z5" >>$LOG
+    echo "  The Synology provided Java may not work with OpenHAB." >>$LOG
     exit 1
   fi
 
   # Is the User Home service enabled?
   UH_SERVICE=$(synogetkeyvalue /etc/synoinfo.conf userHomeEnable)
   if [ "${UH_SERVICE}" == "no" ]; then
-    echo "The User Home service is not enabled. Please enable this feature in the User control panel in DSM."
+    echo "  ERROR:" >>$LOG
+    echo "  The User Home service is not enabled. Please enable this feature in the User control panel in DSM." >>$LOG
     exit 1
   fi
 
-  synoshare --get public > /dev/null || (
-    echo "A shared folder called 'public' could not be found - note this name is case-sensitive. "
-    echo "Please create this using the Shared Folder DSM Control Panel and try again."
+  if [[ ! -d ${SHARE_FOLDER} ]]; then
+    echo "  ERROR:" >>$LOG
+    echo "  A shared folder called '${SHARE_FOLDER}' could not be found - note this name is case-sensitive. " >>$LOG
+    echo "  Please create this using the Shared Folder DSM Control Panel and try again." >>$LOG
     exit 1
-  )
+  fi
 
-  echo "Get new version"
+  echo "  Get new version" >>$LOG
   cd ${TEMP_FOLDER}
   # go through list of files
   for WGET_URL in ${INSTALL_FILES}; do
     WGET_FILENAME="$(echo ${WGET_URL} | sed -r "s%^.*/(.*)%\1%")"
-    echo "Processing ${WGET_FILENAME}"
+    echo "  Processing ${WGET_FILENAME}" >>$LOG
     [ -f "${TEMP_FOLDER}/${WGET_FILENAME}" ] && rm ${TEMP_FOLDER}/${WGET_FILENAME}
     # use local file first
-    if [ -f "${PUBLIC_FOLDER}/${WGET_FILENAME}" ]; then
-      echo "Found file locally - copying"
-      cp ${PUBLIC_FOLDER}/${WGET_FILENAME} ${TEMP_FOLDER}
+    if [ -f "${SHARE_FOLDER}/${WGET_FILENAME}" ]; then
+      echo "  Found file locally - copying" >>$LOG
+      cp ${SHARE_FOLDER}/${WGET_FILENAME} ${TEMP_FOLDER}
     else
       wget -nv --no-check-certificate --output-document=${WGET_FILENAME} ${WGET_URL}
       if [[ $? != 0 ]]; then
-          echo "There was a problem downloading ${WGET_FILENAME} from the download link:"
-          echo "'${WGET_URL}'"
-          echo "Alternatively, download this file manually and place it in the 'public' shared folder and start installation again."
-          if [ -z "${PUBLIC_FOLDER}" ]; then
-            echo "Note: You must create a 'public' shared folder first on your primary volume"
+          echo "  ERROR:" >>$LOG
+          echo "  There was a problem downloading ${WGET_FILENAME} from the download link:" >>$LOG
+          echo "  '${WGET_URL}'" >>$LOG
+          echo "  Alternatively, download this file manually and place it in the '${SHARE_FOLDER}' shared folder and start installation again." >>$LOG
+          if [ -z "${SHARE_FOLDER}" ]; then
+            echo "  Note: You must create a '${SHARE_FOLDER}' shared folder first on your primary volume" >>$LOG
           fi
           exit 1
       fi
     fi
   done
 
+  echo "done" >>$LOG
   exit 0
 }
 
 
 postinst ()
 {
-  #create daemon user
-  echo "Create '${DAEMON_USER}' daemon user"
+  echo "Start postinst..." >>$LOG
+  #create daemon user if not exists
+  echo "  Create '${DAEMON_USER}' daemon user" >>$LOG
   synouser --add ${DAEMON_USER} ${DAEMON_PASS} "${DAEMON_ID}" 0 "" ""
   sleep 3
-
+  
   #add openhab user & handle possible device groups
   synogroup --member dialout ${DAEMON_USER}
   synogroup --member uucp ${DAEMON_USER}
@@ -100,7 +131,7 @@ postinst ()
   su - ${DAEMON_USER} -s /bin/sh -c "echo export OPENHAB_PID=~/.daemon.pid >> .profile"
 
   #extract main archive
-  echo "Install new version"
+  echo "  Install new version" >>$LOG
   cd ${TEMP_FOLDER}
   mkdir ${EXTRACTED_FOLDER}
   tar -xf ${TEMP_FOLDER}/${DOWNLOAD_FILE1} -C ${EXTRACTED_FOLDER} && rm ${TEMP_FOLDER}/${DOWNLOAD_FILE1}
@@ -108,23 +139,31 @@ postinst ()
   rmdir ${TEMP_FOLDER}/${EXTRACTED_FOLDER}
   chmod +x ${SYNOPKG_PKGDEST}/${ENGINE_SCRIPT}
 
-  echo "Create conf/addon links"
+  # if selected create folders for home dir 
+  if [ "${pkgwizard_home_dir}" == "true" ]; then
+    echo "  Create conf/addon folders for home dir" >>$LOG
+    mkdir -p ${OH_CONF}
+    mkdir -p ${OH_ADDONS}
+  fi
+  echo "  Create conf/addon links" >>$LOG
   #if configdir exists in public folder -> create a symbolic link
-  if [ -d ${PUBLIC_CONF} ]; then
-    PUBLIC_OH_FOLDERS_EXISTS=yes
-    mv -u ${SYNOPKG_PKGDEST}/conf/* ${PUBLIC_CONF}
+  if [ -d ${OH_CONF} ]; then
+    echo "    Move conf to ${OH_CONF}" >>$LOG
+    OH_FOLDERS_EXISTS=yes
+    mv -u ${SYNOPKG_PKGDEST}/conf/* ${OH_CONF}
     rm -r ${SYNOPKG_PKGDEST}/conf
-    ln -s ${PUBLIC_CONF} ${SYNOPKG_PKGDEST}
-    synoacltool -get ${PUBLIC_CONF} | grep -F ${DAEMON_ACL} > /dev/null || synoacltool -add ${PUBLIC_CONF} ${DAEMON_ACL}
+    ln -s ${OH_CONF} ${SYNOPKG_PKGDEST}
+    synoacltool -get ${OH_CONF} | grep -F ${DAEMON_ACL} > /dev/null || synoacltool -add ${OH_CONF} ${DAEMON_ACL}
   fi
 
   #if public addons dir exists in public folder -> create a symbolic link
-  if [ -d ${PUBLIC_ADDONS} ]; then
-    PUBLIC_OH_FOLDERS_EXISTS=yes
-    mv -u ${SYNOPKG_PKGDEST}/addons/* ${PUBLIC_ADDONS}
+  if [ -d ${OH_ADDONS} ]; then
+    echo "    Move addons to ${OH_ADDONS}" >>$LOG
+    OH_FOLDERS_EXISTS=yes
+    mv -u ${SYNOPKG_PKGDEST}/addons/* ${OH_ADDONS}
     rm -r ${SYNOPKG_PKGDEST}/addons
-    ln -s ${PUBLIC_ADDONS} ${SYNOPKG_PKGDEST}
-    synoacltool -get ${PUBLIC_ADDONS} | grep -F ${DAEMON_ACL} > /dev/null || synoacltool -add ${PUBLIC_ADDONS} ${DAEMON_ACL}
+    ln -s ${OH_ADDONS} ${SYNOPKG_PKGDEST}
+    synoacltool -get ${OH_ADDONS} | grep -F ${DAEMON_ACL} > /dev/null || synoacltool -add ${OH_ADDONS} ${DAEMON_ACL}
   fi
 
   #add log file
@@ -133,15 +172,24 @@ postinst ()
   
   # Restore UserData if exists
   if [ -d ${BACKUP_FOLDER} ]; then
+    echo "  Restore userdata to ${SYNOPKG_PKGDEST}" >>$LOG
     cp -arf ${BACKUP_FOLDER}/userdata ${SYNOPKG_PKGDEST}/
+    if [ -d ${BACKUP_FOLDER}/userdir ]; then
+      echo "  Restore configuration files to ${OH_FOLDER}" >>$LOG
+      cp -arf ${BACKUP_FOLDER}/userdir/* ${OH_FOLDER}
+    fi 
   fi
   
   #change owner of folder tree
-  echo "Fix permissions"
-  if [ $PUBLIC_OH_FOLDERS_EXISTS == yes ]; then
-    synoshare --setuser public RO + ${DAEMON_USER}
-    chown -hR ${DAEMON_USER} ${PUBLIC_CONF}
-    chown -hR ${DAEMON_USER} ${PUBLIC_ADDONS}
+  echo "  Fix permissions" >>$LOG
+  if [ $OH_FOLDERS_EXISTS == yes ]; then
+    if [ "${pkgwizard_public_std}" == "true" ]; then
+      synoshare --setuser public RO + ${DAEMON_USER}
+    elif [ "${pkgwizard_public_shome}"  == "true" ]; then
+      synoshare --setuser smarthome RO + ${DAEMON_USER}
+    fi
+    chown -hR ${DAEMON_USER} ${OH_CONF}
+    chown -hR ${DAEMON_USER} ${OH_ADDONS}
   fi
   chown -hR ${DAEMON_USER} ${SYNOPKG_PKGDEST}
   chmod -R u+w ${SYNOPKG_PKGDEST}/userdata
@@ -154,6 +202,7 @@ postinst ()
     chmod 777 /dev/ttyACM1
   fi
   
+  echo "done" >>$LOG
   echo "Installation done." > $SYNOPKG_TEMP_LOGFILE;
 
   exit 0
@@ -162,18 +211,21 @@ postinst ()
 
 preuninst ()
 {
+  echo "Start preuninst..." >>$LOG
   #make sure server is stopped
   if su - ${DAEMON_USER} -s /bin/sh -c "cd ${SYNOPKG_PKGDEST}/runtime/karaf/bin && ./stop &"; then
     rm -f $PIDFILE
   fi
   sleep 10
 
+  echo "done" >>$LOG
   exit 0
 }
 
 
 postuninst ()
 {
+  echo "Start postuninst..." >>$LOG
   # Determine folder before deleting daemon
   DAEMON_HOME="$(synouser --get ${DAEMON_USER} | grep -oP 'User Dir.+\[\K[^]]+')"
 
@@ -185,22 +237,33 @@ postuninst ()
   if [ -e "${DAEMON_HOME}" ]; then
     rm -r "${DAEMON_HOME}"
   else
-    echo "Daemon user folder '${DAEMON_HOME}' not found - nothing deleted"
+    echo "  Daemon user folder '${DAEMON_HOME}' not found - nothing deleted" >>$LOG
   fi
   
+  echo "done" >>$LOG
   exit 0
 }
 
 
 preupgrade ()
 {
+  echo "Start preupgrade..." >>$LOG
+  
+  if [[ ! -d ${SHARE_FOLDER} ]]; then
+    echo "  ERROR:" >>$LOG
+    echo "  A shared folder called '${SHARE_FOLDER}' could not be found - note this name is case-sensitive. " >>$LOG
+    echo "  Please create this using the Shared Folder DSM Control Panel and try again." >>$LOG
+    exit 1
+  fi
+  
   #make sure server is stopped
-  echo "Stop server"
+  echo "  Stop server" >>$LOG
   if su - ${DAEMON_USER} -s /bin/sh -c "cd ${SYNOPKG_PKGDEST}/runtime/karaf/bin && ./stop &"; then
     rm -f $PIDFILE
   fi
   sleep 10
   
+  echo "  Remove tmp, cache and runtime dirs" >>$LOG
   # Remove tmp, logs, cache and runtime dirs
   if [ -d ${SYNOPKG_PKGDEST}/userdata/tmp ]; then
   	rm -rf ${SYNOPKG_PKGDEST}/userdata/tmp
@@ -222,6 +285,7 @@ preupgrade ()
     rm -rf ${SYNOPKG_PKGDEST}/runtime
   fi
   
+  echo "  Remove openHAB system files" >>$LOG
   # Remove openHAB system files...
   rm -f ${SYNOPKG_PKGDEST}/userdata/etc/all.policy
   rm -f ${SYNOPKG_PKGDEST}/userdata/etc/branding.properties
@@ -236,19 +300,32 @@ preupgrade ()
   rm -f ${SYNOPKG_PKGDEST}/userdata/etc/org.apache.karaf*
   rm -f ${SYNOPKG_PKGDEST}/userdata/etc/org.ops4j.pax.url.mvn.cfg
   
+  echo "  Create backup" >>$LOG
   # Create backup
   mkdir -p ${BACKUP_FOLDER}/userdata
   mv ${SYNOPKG_PKGDEST}/userdata/* ${BACKUP_FOLDER}/userdata
+  
+  # save home dir content if exists or save current content for the new location
+  LINK_FOLDER="$(readlink ${SYNOPKG_PKGDEST}/conf)"
+  if [[ "${pkgwizard_home_dir}" == "true" || ${LINK_FOLDER} != ${OH_CONF} ]]; then
+    echo "  Save symbolic link folders" >>$LOG
+    LINK_FOLDER="$(dirname ${LINK_FOLDER})"
+    mkdir -p ${BACKUP_FOLDER}/userdir
+    mv ${LINK_FOLDER}/* ${BACKUP_FOLDER}/userdir
+  fi
 
+  echo "done" >>$LOG
   exit 0
 }
 
 
 postupgrade ()
 {
+  echo "Start postupgrade..." >>$LOG
   # Remove all backups after installation
   rm -rf ${SYNOPKG_PKGDEST}-backup*
   
+  echo "done" >>$LOG
   echo "Update done." > $SYNOPKG_TEMP_LOGFILE
   
   exit 0
